calendar now properly reacts to date change
